requirements/base.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
psycopg==3.1.12
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1